Crud promotor.category finish
