+++ src/front/pages/Home.jsx
@@ -70,6 +70,12 @@ export const Home = () => {
 			<Link className="p-2" to="/group">
 				<button className="btn btn-primary">Group CRUD</button>
 			</Link>
+
+			<Link to="/promotor-category/list">
+                <button className="btn btn-primary">
+                    Promotor-Category CRUD
+                </button>
+            </Link>
 		</div>
 	);
 };
