+++ src/front/routes.jsx
@@ -37,6 +37,10 @@ import { CreateGroup } from "./pages/Group/CreateGroup"
 import { EditGroup } from "./pages/Group/EditGroup"
 import { DeleteGroup } from "./pages/Group/DeleteGroup"
 import { GroupDetail } from "./pages/Group/GroupDetail"
+import { CreatePromotorCategory } from "./pages/Promotor-categories/CreatePromotorCategory";
+import { PromotorCategoryList } from "./pages/Promotor-categories/PromotorCategoryList";
+import { EditPromotorCategory } from "./pages/Promotor-categories/EditPromotorCategory";
+
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
@@ -80,6 +84,9 @@ export const router = createBrowserRouter(
       <Route path="/group-edit/:theId" element={<EditGroup />} />
       <Route path="/group-delete/:theId" element={<DeleteGroup />} />
       <Route path="/group/:theId/detail" element={<GroupDetail/>} /> 
+      <Route path="/promotor-category/create" element={<CreatePromotorCategory />} />
+      <Route path="/promotor-category/list" element={<PromotorCategoryList />} />
+      <Route path="/promotor-category/edit/:id" element={<EditPromotorCategory />} />
     </Route>
   )
 );
